Ported to vs 2017 and wrote some optimizations
