+++ PowerMode.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.6
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PowerMode", "PowerMode\PowerMode.csproj", "{AD9EADA6-E3E8-4E7D-934E-24765E47BF1F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PowerModeV2", "PowerModeV2\PowerModeV2.csproj", "{F27430AC-C454-475D-88EC-FFE9124AA835}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AD9EADA6-E3E8-4E7D-934E-24765E47BF1F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AD9EADA6-E3E8-4E7D-934E-24765E47BF1F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AD9EADA6-E3E8-4E7D-934E-24765E47BF1F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AD9EADA6-E3E8-4E7D-934E-24765E47BF1F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F27430AC-C454-475D-88EC-FFE9124AA835}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F27430AC-C454-475D-88EC-FFE9124AA835}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F27430AC-C454-475D-88EC-FFE9124AA835}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F27430AC-C454-475D-88EC-FFE9124AA835}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
