--- PowerMode/packages.config
@@ -1,52 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.VisualStudio.CoreUtility" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Imaging" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.OLE.Interop" version="7.10.6070" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.SDK.VsixSuppression" version="14.1.15" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.14.0" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Immutable.10.0" version="10.0.30319" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Immutable.11.0" version="11.0.50727" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Immutable.12.0" version="12.0.21003" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Immutable.14.0" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Interop" version="7.10.6071" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.10.0" version="10.0.30319" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.11.0" version="11.0.61030" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.12.0" version="12.0.30110" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.8.0" version="8.0.50727" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Shell.Interop.9.0" version="9.0.30729" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Text.Data" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Text.Logic" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Text.UI" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Text.UI.Wpf" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.TextManager.Interop" version="7.10.6070" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.TextManager.Interop.8.0" version="8.0.50727" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Threading" version="14.1.111" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Utilities" version="14.1.24720" targetFramework="net452" />
-  <package id="Microsoft.VisualStudio.Validation" version="14.1.111" targetFramework="net452" />
-  <package id="Microsoft.VSSDK.BuildTools" version="14.1.24720" targetFramework="net452" developmentDependency="true" />
-  <package id="VSSDK.CoreUtility" version="11.0.4" targetFramework="net452" />
-  <package id="VSSDK.CoreUtility.11" version="11.0.4" targetFramework="net452" />
-  <package id="VSSDK.DTE" version="7.0.4" targetFramework="net452" />
-  <package id="VSSDK.GraphModel" version="11.0.4" targetFramework="net452" />
-  <package id="VSSDK.IDE" version="7.0.4" targetFramework="net452" />
-  <package id="VSSDK.IDE.10" version="10.0.4" targetFramework="net452" />
-  <package id="VSSDK.IDE.11" version="11.0.4" targetFramework="net452" />
-  <package id="VSSDK.IDE.12" version="12.0.4" targetFramework="net452" />
-  <package id="VSSDK.IDE.8" version="8.0.4" targetFramework="net452" />
-  <package id="VSSDK.IDE.9" version="9.0.4" targetFramework="net452" />
-  <package id="VSSDK.OLE.Interop" version="7.0.4" targetFramework="net452" />
-  <package id="VSSDK.Shell.12" version="12.0.4" targetFramework="net452" />
-  <package id="VSSDK.Shell.Immutable.10" version="10.0.4" targetFramework="net452" />
-  <package id="VSSDK.Shell.Immutable.11" version="11.0.4" targetFramework="net452" />
-  <package id="VSSDK.Shell.Immutable.12" version="12.0.4" targetFramework="net452" />
-  <package id="VSSDK.Shell.Interop" version="7.0.4" targetFramework="net452" />
-  <package id="VSSDK.Shell.Interop.8" version="8.0.4" targetFramework="net452" />
-  <package id="VSSDK.Shell.Interop.9" version="9.0.4" targetFramework="net452" />
-  <package id="VSSDK.Text" version="11.0.4" targetFramework="net452" />
-  <package id="VSSDK.Text.11" version="11.0.4" targetFramework="net452" />
-  <package id="VSSDK.TextManager.Interop" version="7.0.4" targetFramework="net452" />
-  <package id="VSSDK.TextManager.Interop.8" version="8.0.4" targetFramework="net452" />
-  <package id="VSSDK.Threading" version="12.0.4" targetFramework="net452" />
-  <package id="VSSDK.Threading.12" version="12.0.4" targetFramework="net452" />
-</packages>
