+++ PowerModeV2/ExplosionViewportAdornmentTextViewCreationListener.cs
