+++ PowerModeV2/Extensions/RandomExtensions.cs
@@ -21,7 +21,8 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 
-*/using System;
+*/
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -36,6 +37,7 @@ namespace PowerMode.Extensions
         {
             return random.Next(0, 2) == 1 ? 1 : -1;
         }
+
         public static Color NextColor(this Random random)
         {
             byte[] bytes = new byte[3];
