+++ PowerModeV2/PowerMode/IPowerMode.cs
