+++ PowerModeV2/PowerMode/PowerModeService.cs
