+++ PowerModeV2/VSPackage.resx
@@ -121,7 +121,7 @@
     <value>PowerModeOptionsPackage Extension</value>
   </data>
   <data name="1112" xml:space="preserve">
-    <value>PowerModeOptionsPackage Visual Stuido Extension Detailed Info</value>
+    <value>PowerModeOptionsPackage Visual Studio Extension Detailed Info</value>
   </data>
   <data name="1113" xml:space="preserve">
     <value>General</value>
